# backend/app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bible 365 API",
    "description": "API para aplicação de leitura bíblica diária",
    "version": "1.0.0",
    "contact": {
      "email": "suporte@bible365.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Registra novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "joao.silva"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "joao@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "senha123"
                  }
                },
                "required": ["username", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User registered successfully"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email já registrado"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Login do usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/api/v1/reading-plan/start/{plan_type}": {
      "post": {
        "tags": ["Plano de Leitura"],
        "summary": "Inicia um novo plano de leitura",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "plan_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["GENESIS_TO_REV"]
            },
            "description": "Tipo do plano de leitura"
          }
        ],
        "responses": {
          "201": {
            "description": "Plano criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Plano criado com sucesso!"
                    },
                    "plan_id": {
                      "type": "integer",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Usuário já possui um plano ativo ou tipo de plano inválido"
          }
        }
      }
    },
    "/api/v1/reading-plan/current": {
      "get": {
        "tags": ["Plano de Leitura"],
        "summary": "Obtém o plano atual do usuário",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do plano atual",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plan_id": {
                      "type": "integer"
                    },
                    "current_day": {
                      "type": "integer"
                    },
                    "streak": {
                      "type": "integer"
                    },
                    "start_date": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "last_read_date": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "progress": {
                      "type": "object",
                      "properties": {
                        "total_days": {
                          "type": "integer"
                        },
                        "completed_readings": {
                          "type": "integer"
                        },
                        "total_readings": {
                          "type": "integer"
                        },
                        "percentage": {
                          "type": "number"
                        }
                      }
                    },
                    "todays_readings": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "book": {
                            "type": "string"
                          },
                          "chapter": {
                            "type": "integer"
                          },
                          "completed": {
                            "type": "boolean"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Nenhum plano ativo encontrado"
          }
        }
      }
    },
    "/api/v1/reading-plan/complete/{chapter_id}": {
      "post": {
        "tags": ["Plano de Leitura"],
        "summary": "Marca uma leitura como concluída",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da leitura a ser marcada como concluída"
          }
        ],
        "responses": {
          "200": {
            "description": "Leitura marcada como concluída",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "progress": {
                      "type": "object",
                      "properties": {
                        "total_days": {
                          "type": "integer"
                        },
                        "completed_readings": {
                          "type": "integer"
                        },
                        "total_readings": {
                          "type": "integer"
                        },
                        "percentage": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Leitura não encontrada ou já concluída"
          }
        }
      }
    },
    "/api/v1/reading-plan/history": {
      "get": {
        "tags": ["Plano de Leitura"],
        "summary": "Obtém histórico de leituras do usuário",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Número da página"
          },
          {
            "name": "per_page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Itens por página"
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico de leituras",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "readings": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "book": {
                            "type": "string"
                          },
                          "chapter": {
                            "type": "integer"
                          },
                          "completed_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "pages": {
                      "type": "integer"
                    },
                    "current_page": {
                      "type": "integer"
                    },
                    "per_page": {
                      "type": "integer"
                    },
                    "has_next": {
                      "type": "boolean"
                    },
                    "has_prev": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}